+++ libjankscripten/src/javascript/mod.rs
@@ -13,13 +13,13 @@ mod desugar_updates;
 mod desugar_vardecls;
 mod lift_vars;
 pub mod name_gen;
-mod swc_parser;
+mod parser;
 pub mod syntax;
 pub mod walk;
 
 pub use desugar::*;
 pub use name_gen::*;
-pub use swc_parser::*;
+pub use parser::*;
 pub use syntax::*;
 pub use walk::*;
